# docs/swagger.yaml
basePath: /
definitions:
...
  report.ReportRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
  report.UpdateReportRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  response.ErrorResponse:
    description: Error response
    properties:
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.SuccessResponse:
    description: Success response
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:5000
info:
  contact:
    email: support@go-sosmed.com
    name: API Support
  description: API for Golang Social Media Application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GO-SOSMED API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - User
...
  /api/posts/{post_id}/reports:
    post:
      consumes:
      - application/json
      description: Report a post for inappropriate content
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Report data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/report.ReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a report
      tags:
      - Report
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: Registration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register user
      tags:
      - User
  /api/reports:
    get:
      consumes:
      - application/json
      description: Retrieve all reports (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all reports
      tags:
      - Report
  /api/reports/{report_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific report by ID (Admin only)
      parameters:
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get report by ID
      tags:
      - Report
  /api/reports/{report_id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a report (Admin only)
      parameters:
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: integer
      - description: Status update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/report.UpdateReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update report status
      tags:
      - Report
  /api/users/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve user information by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - User
...
  /api/users/me:
    put:
      consumes:
      - multipart/form-data
      description: Update authenticated user's profile information
      parameters:
      - description: Username
        in: formData
        name: username
        type: string
      - description: User bio
        in: formData
        name: bio
        type: string
      - description: Avatar image
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"